# crdb: better seek, fix untyped versions bug