@RaduBerinde 2024-10-24
I plan to address the TODO to add tests with mixed suffix types in a separate PR (I'll also add some randomized tests).